requirements.txt
openai
sounddevice
numpy
webrtcvad-wheels
python-dotenv
fastapi
uvicorn
